Migrate Elasticsearch to IndexerInterface and SearcherInterface
