+++ packages/seal-elasticsearch-adapter/ElasticsearchAdapter.php
@@ -4,22 +4,27 @@ namespace Schranz\Search\SEAL\Adapter\Elasticsearch;
 
 use Elastic\Elasticsearch\Client;
 use Schranz\Search\SEAL\Adapter\AdapterInterface;
-use Schranz\Search\SEAL\Adapter\ConnectionInterface;
+use Schranz\Search\SEAL\Adapter\IndexerInterface;
 use Schranz\Search\SEAL\Adapter\SchemaManagerInterface;
+use Schranz\Search\SEAL\Adapter\SearcherInterface;
 
 final class ElasticsearchAdapter implements AdapterInterface
 {
-    private readonly ConnectionInterface $connection;
-
     private readonly SchemaManagerInterface $schemaManager;
+
+    private readonly IndexerInterface $indexer;
+
+    private readonly SearcherInterface $searcher;
 
     public function __construct(
         private readonly Client $client,
-        ?ConnectionInterface $connection = null,
         ?SchemaManagerInterface $schemaManager = null,
+        ?IndexerInterface $indexer = null,
+        ?SearcherInterface $searcher = null,
     ) {
-        $this->connection = $connection ?? new ElasticsearchConnection($client);
         $this->schemaManager = $schemaManager ?? new ElasticsearchSchemaManager($client);
+        $this->indexer = $indexer ?? new ElasticsearchIndexer($client);
+        $this->searcher = $searcher ?? new ElasticsearchSearcher($client);
     }
 
     public function getSchemaManager(): SchemaManagerInterface
@@ -27,8 +32,13 @@ final class ElasticsearchAdapter implements AdapterInterface
         return $this->schemaManager;
     }
 
-    public function getConnection(): ConnectionInterface
+    public function getIndexer(): IndexerInterface
     {
-        return $this->connection;
+        return $this->indexer;
+    }
+
+    public function getSearcher(): SearcherInterface
+    {
+        return $this->searcher;
     }
 }
